Something is not right with the momentum (#1)
After making the correction to the momentum (not scaling it by the
learning rate), we have some odd effects on the learning when the
momentum is non-zero. At zero momentum, we get training performance
that seems to make sense, but now, it seems weird. In fact, for
non-zero momentums, we don’t get any kind of reasonable training.