Correcting hidden sampling (#15)
To be more in line with Hinton and the theory, lets make sure that that
we are using the conditional probabilities for the hidden units when we
calculate the difference in the positive and negative phases.

So, technically, now, the `gibbs` code is no longer a gibbs sampler,
per se, but is returning the means for the hidden units. Which is just
as well, since this is what we want for the contrastive-divergence
computation, anyway.

Thanks @marylou-gabrie and @AurelienDecelle