Restructure: New `RBMBase.jl`
The goal is to split up some of this functionality between files in
order to make the code more understandable. To do this, with the linear
fashion in which all the calls to the “headers” are included, I need to
put the base structure definitions for the RBM in their own file,
subsequently followed by more complicated code.