Adding save_progress functionality for DBMs
When joint training, so with the function `fit`, one can add the
`save_progress` kwarg with identical syntax to RBMs.
* The io.jl file has new methods for DBM `save_params` and
`append_params` (although on this branch there are not yet methods for
RBMs). The current version saves all fields of the DBMs. This is a bit
overkill, and we might want revise it.
* Training.jl has the `fit` function updated, saving the model if
save_progress is different from the default value `nothing`.